+++ tm2/pkg/db/types.go
@@ -1,9 +1,5 @@
 package db
 
-import (
-	corestore "cosmossdk.io/core/store"
-)
-
 // DBs are goroutine safe.
 type DB interface {
 	// Get returns nil iff key doesn't exist.
@@ -64,8 +60,35 @@ type DB interface {
 // ----------------------------------------
 // Batch
 
+// Batch represents a group of writes. They may or may not be written atomically depending on the
+// backend. Callers must call Close on the batch when done.
+//
 // Batch Close must be called when the program no longer needs the object.
-type Batch = corestore.Batch
+type Batch interface {
+	// Set sets a key/value pair.
+	// CONTRACT: key, value readonly []byte
+	Set(key, value []byte) error
+
+	// Delete deletes a key/value pair.
+	// CONTRACT: key readonly []byte
+	Delete(key []byte) error
+
+	// Write writes the batch, possibly without flushing to disk. Only Close() can be called after,
+	// other methods will error.
+	Write() error
+
+	// WriteSync writes the batch and flushes it to disk. Only Close() can be called after, other
+	// methods will error.
+	WriteSync() error
+
+	// Close closes the batch. It is idempotent, but calls to other methods afterwards will error.
+	Close() error
+
+	// GetByteSize that returns the current size of the batch in bytes. Depending on the implementation,
+	// this may return the size of the underlying LSM batch, including the size of additional metadata
+	// on top of the expected key and value total byte count.
+	GetByteSize() (int, error)
+}
 
 // ----------------------------------------
 // Iterator
@@ -81,4 +104,37 @@ defer itr.Close()
 		// ...
 	}
 */
-type Iterator = corestore.Iterator
+// Iterator represents an iterator over a domain of keys. Callers must call
+// Close when done. No writes can happen to a domain while there exists an
+// iterator over it. Some backends may take out database locks to ensure this
+// will not happen.
+//
+// Callers must make sure the iterator is valid before calling any methods on it,
+// otherwise these methods will panic.
+type Iterator interface {
+	// Domain returns the start (inclusive) and end (exclusive) limits of the iterator.
+	Domain() (start, end []byte)
+
+	// Valid returns whether the current iterator is valid. Once invalid, the Iterator remains
+	// invalid forever.
+	Valid() bool
+
+	// Next moves the iterator to the next key in the database, as defined by order of iteration.
+	// If Valid returns false, this method will panic.
+	Next()
+
+	// Key returns the key at the current position. Panics if the iterator is invalid.
+	// Note, the key returned should be a copy and thus safe for modification.
+	Key() []byte
+
+	// Value returns the value at the current position. Panics if the iterator is
+	// invalid.
+	// Note, the value returned should be a copy and thus safe for modification.
+	Value() []byte
+
+	// Error returns the last error encountered by the iterator, if any.
+	Error() error
+
+	// Close closes the iterator, releasing any allocated resources.
+	Close() error
+}