+++ tm2/pkg/store/iavl/store.go
@@ -5,16 +5,15 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/cosmos/iavl"
 	ics23 "github.com/cosmos/ics23/go"
 
+	"github.com/gnolang/gno/tm2/pkg/iavl"
 	"github.com/gnolang/gno/tm2/pkg/amino"
 	abci "github.com/gnolang/gno/tm2/pkg/bft/abci/types"
 	"github.com/gnolang/gno/tm2/pkg/crypto/merkle"
 	dbm "github.com/gnolang/gno/tm2/pkg/db"
 	"github.com/gnolang/gno/tm2/pkg/errors"
 	"github.com/gnolang/gno/tm2/pkg/std"
-
 	"github.com/gnolang/gno/tm2/pkg/store/cache"
 	serrors "github.com/gnolang/gno/tm2/pkg/store/errors"
 	"github.com/gnolang/gno/tm2/pkg/store/types"