+++ tm2/pkg/store/prefix/store_test.go
@@ -6,12 +6,11 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	cosmosiavl "github.com/cosmos/iavl"
-
+	"github.com/gnolang/gno/tm2/pkg/iavl"
 	"github.com/gnolang/gno/tm2/pkg/db/memdb"
 	"github.com/gnolang/gno/tm2/pkg/store/dbadapter"
 	"github.com/gnolang/gno/tm2/pkg/store/gas"
-	"github.com/gnolang/gno/tm2/pkg/store/iavl"
+	storeiavl "github.com/gnolang/gno/tm2/pkg/store/iavl"
 	"github.com/gnolang/gno/tm2/pkg/store/types"
 )
 
@@ -91,8 +90,8 @@ func TestIAVLStorePrefix(t *testing.T) {
 	t.Parallel()
 
 	db := memdb.NewMemDB()
-	tree := cosmosiavl.NewMutableTree(db, cacheSize, false, cosmosiavl.NewNopLogger())
-	iavlStore := iavl.UnsafeNewStore(tree, types.StoreOptions{
+	tree := iavl.NewMutableTree(db, cacheSize, false, iavl.NewNopLogger())
+	iavlStore := storeiavl.UnsafeNewStore(tree, types.StoreOptions{
 		PruningOptions: types.PruningOptions{
 			KeepRecent: numRecent,
 			KeepEvery:  storeEvery,